requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
jinja2==3.1.4
python-multipart==0.0.9
python-dotenv==1.0.1
itsdangerous==2.2.0
bcrypt==4.2.0
gunicorn==22.0.0
